+++ packages/plugin-react-app/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 2.2.1
 
 - [feat] define variables of `CDN_PATH`
+- [fix] bump version of `@builder/pack`(^0.6.0)
+- [fix] remove `webpackHotDevClient`
+- [fix] `vitePlugins` only take effect when `vite` is true
 
 ## 2.2.0